Simplify asprintf dependency, add Unix platform
