+++ Makefile
@@ -1,7 +1,7 @@
 CC=$(PREFIX)gcc
 STRIP=$(PREFIX)strip
 
-SRC=pspack.c index.c asprintf.c vasprintf-msvc-nofuture.c minilzo.c util.c fs.c 
+SRC=pspack.c index.c asprintf.c minilzo.c util.c fs.c 
 EXE=pspack.exe
 CFLAGS=-Wall -O2
