+++ asprintf.h
@@ -25,8 +25,6 @@ extern "C"
 {
 #endif
 
-#define insane_free(ptr) { free(ptr); ptr = 0; }
-
 int vasprintf(char **strp, const char *fmt, va_list ap);
 int asprintf(char **strp, const char *fmt, ...);
